refactored probing script
The probing is done rolling out all the token representation for each layer.
To each token representation the token's target are associated (merging the subtoken with average or deletion),
a token-wise dataset is crafted for each token.
A cross-validation (k_fold parametrized) is used to compute a test metric (100-MAE).